Added a new parameter

+++ source/c++/include/config.hpp
@@ -20,7 +20,7 @@ namespace core{
     extern size_t _velocityDeadBand;
     extern float32 _minScore;
     extern uint8_t _moveToLastSeen;
-    extern uint8_t _minConfidence;
+    extern float32 _minConfidence;
 } 
 
 namespace visualization{
